+++ test/integration/elastic_db_repo/create_repo.py
@@ -30,6 +30,43 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        get_val
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.args_array = dict()
+
+    def get_val(self, skey, def_val=None):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest
@@ -54,6 +91,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.args = ArgParser()
         self.base_dir = "test/integration/elastic_db_repo"
         self.test_path = os.path.join(os.getcwd(), self.base_dir)
         self.config_path = os.path.join(self.test_path, "config")
@@ -85,10 +123,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-C": self.repo_name, "-l": self.cfg.log_repo_dir}
+        self.args.args_array = {
+            "-C": self.repo_name, "-l": self.cfg.log_repo_dir}
 
-        self.assertFalse(elastic_db_repo.create_repo(self.els,
-                                                     args_array=args_array))
+        self.assertFalse(elastic_db_repo.create_repo(self.els, args=self.args))
 
     def test_createrepo_arg(self):
 
@@ -100,9 +138,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.assertFalse(elastic_db_repo.create_repo(
-            self.els, repo_name=self.repo_name, repo_dir=self.cfg.log_repo_dir,
-            args_array={}))
+        self.assertFalse(
+            elastic_db_repo.create_repo(
+                self.els, repo_name=self.repo_name,
+                repo_dir=self.cfg.log_repo_dir, args=self.args))
 
     def tearDown(self):