+++ test/integration/elastic_db_repo/delete_dump.py
@@ -30,6 +30,43 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        get_val
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.args_array = dict()
+
+    def get_val(self, skey, def_val=None):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest
@@ -54,6 +91,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        self.args = ArgParser()
         self.base_dir = "test/integration/elastic_db_repo"
         self.test_path = os.path.join(os.getcwd(), self.base_dir)
         self.config_path = os.path.join(self.test_path, "config")
@@ -98,10 +136,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        args_array = {"-r": self.repo_name, "-S": self.dump_name}
+        self.args.args_array = {"-r": self.repo_name, "-S": self.dump_name}
 
-        self.assertFalse(elastic_db_repo.delete_dump(
-            self.elr, args_array=args_array))
+        self.assertFalse(elastic_db_repo.delete_dump(self.elr, args=self.args))
 
     def test_deletedmp_arg(self):
 
@@ -113,9 +150,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.assertFalse(elastic_db_repo.delete_dump(
-            self.elr, repo_name=self.repo_name, dump_name=self.dump_name,
-            args_array={}))
+        self.assertFalse(
+            elastic_db_repo.delete_dump(
+                self.elr, repo_name=self.repo_name, dump_name=self.dump_name,
+                args=self.args))
 
     def tearDown(self):